+++ packages/components/modules/comments/CommentItem/useCommentOptions/types.ts
@@ -2,6 +2,5 @@ import { CommentItem_comment$data } from '../../../../__generated__/CommentItem_
 
 export interface UseCommentOptionsParams {
   comment?: CommentItem_comment$data
-  onLongPressLeave: () => void
   onEdit: () => void
 }