+++ packages/components/modules/profiles/Members/MemberItem/styled.tsx
@@ -1,6 +1,6 @@
 import { Box, Select as MUISelect, alpha, styled } from '@mui/material'
 
-import { IMemberPersonalInformation } from './types'
+import { MemberPersonalInformationProps } from './types'
 
 export const MemberItemContainer = styled(Box)(({ theme }) => ({
   display: 'flex',
@@ -12,7 +12,7 @@ export const MemberItemContainer = styled(Box)(({ theme }) => ({
 
 export const MemberPersonalInformation = styled(Box, {
   shouldForwardProp: (prop) => prop !== 'isActive',
-})<IMemberPersonalInformation>(({ isActive, theme }) => ({
+})<MemberPersonalInformationProps>(({ isActive, theme }) => ({
   opacity: isActive ? 1 : 0.6,
   display: 'flex',
   gap: theme.spacing(1.5),