+++ packages/components/modules/profiles/Members/MembersList/index.tsx
@@ -2,18 +2,21 @@ import { FC, useMemo } from 'react'
 
 import { LoadingState as DefaultLoadingState } from '@baseapp-frontend/design-system'
 
-import { Box, Divider, Typography } from '@mui/material'
+import { Box, Typography } from '@mui/material'
 import { usePaginationFragment } from 'react-relay'
 import { Virtuoso } from 'react-virtuoso'
 
+import { MemberItemFragment$key } from '../../../../__generated__/MemberItemFragment.graphql'
 import { UserMembersListFragment } from '../../graphql/queries/UserMembersList'
 import DefaultMemberItem from '../MemberItem'
+import MemberListItem from '../MemberListItem'
 import { MemberStatuses, NUMBER_OF_MEMBERS_TO_LOAD_NEXT } from '../constants'
 import { MemberListProps } from '../types'
 
 const MembersList: FC<MemberListProps> = ({
   userRef,
   MemberItem = DefaultMemberItem,
+  MemberItemProps = {},
   LoadingState = DefaultLoadingState,
   LoadingStateProps = {},
   membersContainerHeight = 400,
@@ -40,52 +43,16 @@ const MembersList: FC<MemberListProps> = ({
     )
   }
 
-  const renderMemberItem = (member: any, index: number) => {
-    if (!member) return null
-    if (
-      (member.status === MemberStatuses.active &&
-        members[index - 1]?.status !== MemberStatuses.active) ||
-      (member.status === MemberStatuses.active && !members[index - 1]?.status)
-    ) {
-      return (
-        <>
-          <Divider />
-          <MemberItem member={data} memberRole="owner" status={MemberStatuses.active} />
-          <MemberItem
-            member={member?.user?.profile}
-            memberRole={member?.role}
-            status={member?.status}
-            canChangeMember={data?.canChangeRole || false}
-            userId={member?.user?.id}
-          />
-        </>
-      )
-    }
-    if (member.status !== MemberStatuses.active && !members[index + 1]?.status) {
-      return (
-        <>
-          <MemberItem
-            member={member?.user?.profile}
-            memberRole={member?.role}
-            status={member?.status}
-            userId={member?.user?.id}
-            canChangeMember={data?.canChangeRole || false}
-          />
-          <Divider />
-          <MemberItem member={data} memberRole="owner" status={MemberStatuses.active} />
-        </>
-      )
-    }
-    return (
-      <MemberItem
-        member={member?.user?.profile}
-        memberRole={member?.role}
-        status={member?.status}
-        canChangeMember={data?.canChangeRole || false}
-        userId={member?.user?.id}
-      />
-    )
-  }
+  const renderMemberItem = (member: MemberItemFragment$key, index: number) => (
+    <MemberListItem
+      member={member}
+      data={data}
+      prevMember={members[index - 1]}
+      nextMember={members[index + 1]}
+      MemberItemComponent={MemberItem}
+      memberItemComponentProps={MemberItemProps}
+    />
+  )
 
   if (members.length === 0) {
     return (
@@ -100,16 +67,13 @@ const MembersList: FC<MemberListProps> = ({
 
   return (
     <>
-      {data.members?.totalCount && (
-        <Typography variant="subtitle2" mb={4}>
-          {(data.members?.totalCount ?? 0) + 1} members
-        </Typography>
-      )}
+      <Typography variant="subtitle2" mb={4}>
+        {(data.members?.totalCount ?? 0) + 1} members
+      </Typography>
       <Virtuoso
         style={{ height: membersContainerHeight }}
         data={members}
-        overscan={NUMBER_OF_MEMBERS_TO_LOAD_NEXT}
-        itemContent={(_index, member) => renderMemberItem(member, _index)}
+        itemContent={(_index, member) => member && renderMemberItem(member, _index)}
         components={{
           Footer: renderLoadingState,
         }}