+++ packages/components/modules/profiles/graphql/queries/MemberItem.ts
@@ -1,0 +1,15 @@
+import { graphql } from 'react-relay'
+
+export const MemberItemFragment = graphql`
+  fragment MemberItemFragment on ProfileUserRole {
+    id
+    user {
+      profile {
+        ...ProfileItemFragment
+      }
+      id
+    }
+    role
+    status
+  }
+`